% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doobi1124-package.R
\docType{package}
\name{doobi1124-package}
\alias{doobi1124}
\alias{doobi1124-package}
\title{doobi1124: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
