% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GimmeMyOmics-package.R
\docType{package}
\name{GimmeMyOmics-package}
\alias{GimmeMyOmics}
\alias{GimmeMyOmics-package}
\title{GimmeMyOmics: Omics Data Utilities}
\description{
A comprehensive toolkit for high-throughput omics analysis. It provides functions for differential expression analysis, functional enrichment, and network-based annotation. The package also includes a variety of visualization tools, such as volcano plots, heatmaps, violin plots, and pathway diagrams, to facilitate data exploration and interpretation.
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
