% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{map_enrich}
\alias{map_enrich}
\title{Enrichment map visualization}
\usage{
map_enrich(
  x,
  regex = NULL,
  width = 20,
  FDR = 0.05,
  cex = 1,
  nWords = 2,
  size_edge = 0.5,
  showCategory = 15,
  colour = c(palette_discrete()[1], "gray50", palette_discrete()[2]),
  method = "clusterProfiler",
  ...
)
}
\arguments{
\item{x}{Enrichment result.}

\item{regex}{Regular expression pattern to filter by pathway names.}

\item{width}{Integer for maximum width for pathway name wrapping.}

\item{FDR}{Double for False Discovery Rate threshold for filtering results.}

\item{cex}{Double for the scaling factor for all elements.}

\item{nWords}{Numeric, the number of words in the cluster tags, the default value is 4.}

\item{size_edge}{relative size of edge width}

\item{showCategory}{A number or a vector of terms. If it is a number,
the first n terms will be displayed. If it is a vector of terms,
the selected terms will be displayed.}

\item{colour}{Color palette for gradient coloring.}

\item{method}{A character string specifying the package of the enrichment analysis method.
Options are:
\itemize{
\item \code{enrichR};
\item \code{clusterProfiler}.
}}

\item{...}{Additional arguments passed to \code{enrichplot::emapplot}.}
}
\value{
A ggplot object showing the enrichment map.
}
\description{
Wrapper around \code{emapplot} for creating enrichment maps from
enrichment analysis results. Provides customization options for clustering,
node appearance, and improved aesthetics.
}
\details{
This function extends \code{enrichplot::emapplot} with:
\itemize{
\item Automatic filtering by FDR and term description patterns
\item Enhanced label formatting (wrapping, truncation)
\item Customizable clustering and scaling parameters
}
}
\examples{
\dontrun{
data(geneList, package = "DOSE")
de <- names(geneList)[seq(100)]
ora_res <- enrichGO(de, OrgDb = "org.Hs.eg.db", ont = "BP")

# Basic usage
map_enrich(ora_res)

# With custom parameters
map_enrich(
  ora_res,
  regex = "response",
  FDR = 0.01,
  cex = 1.2,
  showCategory = 20,
  colour = c("blue", "white", "red"),
  cluster.params = list(cluster = FALSE)
)
}

}
\seealso{
\code{\link[enrichplot]{emapplot}}, \code{\link[enrichplot]{pairwise_termsim}}
}
