% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bulk.R
\name{ncbi_description}
\alias{ncbi_description}
\title{Fetches Gene Description from NCBI}
\usage{
ncbi_description(x, metadata_genes, database = "ensembl_gene_id")
}
\arguments{
\item{x}{Character, a gene identifier (either Ensembl or Entrez).}

\item{metadata_genes}{A \code{data.frame} or \code{tibble} containing gene metadata with at least \code{ensembl_gene_id} and \code{gene_name} columns.}

\item{database}{Character, specifying the type of gene identifier provided in \code{x}. Accepted values are \code{"ensembl_gene_id"} or \code{"entrezgene_id"}.}
}
\value{
A character string containing the formatted gene description, or \code{NA} if no valid gene ID is found.
}
\description{
Fetches and processes the gene description from NCBI using either an Ensembl or Entrez gene ID.
}
\details{
The function performs the following steps:
\itemize{
\item Constructs the NCBI gene URL and retrieves the gene summary from the webpage.
\item Cleans the retrieved text by removing redundant phrases and formatting inconsistencies.
\item Returns a formatted gene description prefixed by the gene symbol.
}
}
\examples{
metadata_genes <- tibble(
  ensembl_gene_id = c("ENSG00000141510", "ENSG00000272398"),
  entrezgene_id = c(7157, NA),
  gene_name = c("TP53", "Unknown Gene")
)

ncbi_description("ENSG00000141510", metadata_genes)
ncbi_description("7157", metadata_genes, database = "entrezgene_id")

}
