% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrich.R
\name{print_enrich}
\alias{print_enrich}
\title{Displays Enrichment}
\usage{
print_enrich(x, method = "ora", regex = NULL, pval = 0.05)
}
\arguments{
\item{x}{An enrichment results.}

\item{method}{A character string specifying the enrichment analysis method. Options are:
\itemize{
\item \code{"enrichr"}: Enrichr results (\strong{requires \code{path2gene}}, from \code{enrichR} package).
\item \code{"gsea"}: Gene Set Enrichment Analysis (GSEA) results (from \code{clusterProfiler} package).
\item \code{"ora"}: Over-enrichment Analysis (GSEA) results (from \code{clusterProfiler} package).
}}

\item{regex}{A character vector of regular expressions to filter terms by their descriptions.}

\item{pval}{A numeric threshold for filtering results based on FDR-adjusted p-values.}
}
\value{
A formatted \code{tibble} containing processed enrichment results,
with columns:
\itemize{
\item \code{Description}: The pathway or term name.
\item \code{FDR}: The adjusted p-value.
\item \verb{Nb DEG} / \verb{Nb genes}: Number of differentially expressed genes and total genes in the pathway.
\item \code{DEG/Genes}: Percentage of DEGs in the pathway.
\item \code{Genes}: List of genes involved.
\item Additional columns depending on the enrichment method (e.g., \code{NES}, \code{ID}).
}
}
\description{
Formats and filters enrichment analysis results from enrichment analysis libraries:
\code{enrichR} and \code{clusterProfiler}.
}
\details{
This function standardizes different enrichment result formats, applying transformations such as:
\itemize{
\item Computing the percentage of differentially expressed genes (DEGs) in each pathway.
\item Formatting gene lists for better readability.
\item Filtering results based on FDR and optional regex pattern matching.
}
}
\examples{
# Example 1: Enrichr results
path_name <- c("Neutrophil degranulation", "Macrophage migration")
pval <- c(2.7e-02, 2.89e-03)
genes <- c("ITGB2/ANXA3/STXBP2/SPI1/ITGAM/CD177", "MAPK3/AKIRIN1/CX3CR1/CNN2/LGALS3/B4GALT1/C3AR1")
ids <- c("GO:0043312", "GO:1905517")

enrichr_results <- data.frame(
  Term = path_name,
  Adjusted.P.value = pval,
  Overlap = c("6/12", "7/17"),
  ID = ids,
  Genes = gsub("/", ";", genes)
)
print_enrich(enrichr_results, method = "enrichr")

# Example 2: GSEA results (from clusterProfiler)
gsea_results <- data.frame(
  Description = path_name,
  p.adjust = pval,
  core_enrichment = genes,
  setSize = c(12, 17),
  NES = c(1.75, 2.01),
  ID = ids
)
print_enrich(gsea_results, method = "gsea")

# Example 3: Over-representation enrichment results (from clusterProfiler)
ora_results <- data.frame(
  Description = path_name,
  p.adjust = pval,
  GeneRatio = c("6/12", "7/17"),
  BgRatio = c("60/120", "70/140"),
  geneID = genes,
  ID = ids
)
print_enrich(ora_results)

}
