% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrich.R
\name{str_trunc1}
\alias{str_trunc1}
\title{Cuts a sentence to a given number of characters and leaves the words as integers}
\usage{
str_trunc1(x, n = 20, w = " ")
}
\description{
Cuts a sentence to a given number of characters and leaves the words as integers
}
