% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GimmeMyOmics-package.R, R/doobi1124-package.R
\docType{package}
\name{GimmeMyOmics-package}
\alias{GimmeMyOmics-package}
\title{GimmeMyOmics: Omics data analysis}
\description{
GimmeMyOmics provides specialized tools for analyzing bulk RNA-seq, single-cell RNA-seq, and ATAC-seq (peak calling, chromatin accessibility analysis). It includes methods for differential expression, gene set enrichment, over-representation analysis for pathway enrichment, network analysis (gene co-expression, pathway graph) and single-cell workflows (clustering, trajectory inference, cell-type annotation). It also provides a set of tools for data visualization (e.g. volcano plots, heatmaps, UMAP plots and violin plots). The package is built on top of popular R packages (e.g. Seurat, DESeq2, clusterProfiler).

GimmeMyOmics provides specialized tools for analyzing bulk RNA-seq, single-cell RNA-seq, and ATAC-seq (peak calling, chromatin accessibility analysis). It includes methods for differential expression, gene set enrichment, over-representation analysis for pathway enrichment, network analysis (gene co-expression, pathway graph) and single-cell workflows (clustering, trajectory inference, cell-type annotation). It also provides a set of tools for data visualization (e.g. volcano plots, heatmaps, UMAP plots and violin plots). The package is built on top of popular R packages (e.g. Seurat, DESeq2, clusterProfiler).
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
