% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GimmeMyOmics-package.R
\docType{package}
\name{GimmeMyOmics-package}
\alias{GimmeMyOmics}
\alias{GimmeMyOmics-package}
\title{GimmeMyOmics: Omics data analysis}
\description{
GimmeMyOmics is an integrated R package for high-throughput omics analysis, covering bulk RNA-seq, single-cell RNA-seq, and ATAC-seq. It provides streamlined workflows for differential expression analysis, functional enrichment, and network-based annotation. Designed for both bulk and single-cell data, it supports clustering, trajectory inference, and cell-type annotation while enabling chromatin accessibility studies through peak calling and regulatory element analysis. The package also offers quality control and preprocessing tools, converting kallisto.h5 outputs into transcript- and gene-level count matrices, assessing alignment quality, and extracting fastqc and fastp metrics with detailed visual summaries. Its visualization suite generates publication-ready volcano plots, heatmaps, UMAP projections, violin plots, and pathway graphs, facilitating data exploration. Built on standard R packages like Seurat, DESeq2, and clusterProfiler, GimmeMyOmics simplifies multi-omics workflows, making complex bioinformatics analyses more efficient, reproducible, and accessible.
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
