% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GimmeMyOmics-package.R, R/doobi1124-package.R
\docType{package}
\name{GimmeMyOmics-package}
\alias{GimmeMyOmics-package}
\title{GimmeMyOmics: Omics data analysis}
\description{
GimmeMyOmics analyzes high-throughput omics data, including bulk RNA-seq, single-cell RNA-seq, and ATAC-seq. It provides tools for differential expression analysis, gene set enrichment, and over-representation analysis to uncover functional pathways and regulatory networks. The package supports single-cell workflows such as clustering, trajectory inference, and cell-type annotation, as well as chromatin accessibility analysis, including peak calling and functional interpretation of ATAC-seq data. It also offers quality control and preprocessing features, including conversion of kallisto.h5 outputs into transcript- and gene-level count matrices, assessment of alignment quality, and extraction of fastqc quality metrics with visual summaries. Additionally, it includes a suite of visualization tools for generating volcano plots, heatmaps, UMAP projections, violin plots, and pathway graphs, facilitating intuitive data exploration. The package is built on top of widely used R packages like Seurat, DESeq2, and clusterProfiler.

GimmeMyOmics analyzes high-throughput omics data, including bulk RNA-seq, single-cell RNA-seq, and ATAC-seq. It provides tools for differential expression analysis, gene set enrichment, and over-representation analysis to uncover functional pathways and regulatory networks. The package supports single-cell workflows such as clustering, trajectory inference, and cell-type annotation, as well as chromatin accessibility analysis, including peak calling and functional interpretation of ATAC-seq data. It also offers quality control and preprocessing features, including conversion of kallisto.h5 outputs into transcript- and gene-level count matrices, assessment of alignment quality, and extraction of fastqc quality metrics with visual summaries. Additionally, it includes a suite of visualization tools for generating volcano plots, heatmaps, UMAP projections, violin plots, and pathway graphs, facilitating intuitive data exploration. The package is built on top of widely used R packages like Seurat, DESeq2, and clusterProfiler.
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
